% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghdx_micro_checks.R
\name{is_consistent_type}
\alias{is_consistent_type}
\title{Checking that there are never multiple question types for a single variable}
\usage{
is_consistent_type(book)
}
\arguments{
\item{book}{Checking that there are never multiple question types for a single variable}
}
\value{
TRUE if question type is consistent. False if otherwise. Also prints messages to the log.
}
\description{
Checking that there are never multiple question types for a single variable
}
