% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghdx_micro_checks.R
\name{is_vartype_ever_na}
\alias{is_vartype_ever_na}
\title{This function checks if there is ever an N/A value for the specified variable type}
\usage{
is_vartype_ever_na(data, vartype)
}
\arguments{
\item{data}{Checking that there are never multiple question types for a single variable}

\item{vartype}{A text string, enclosed in "", that denotes the variable type to be checked}
}
\value{
TRUE if value and answer label are never "N/A". False if otherwise. Also prints messages to the log.
}
\description{
This function checks if there is ever an N/A value for the specified variable type
}
