% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghdx_micro_checks.R
\name{missing_from_book}
\alias{missing_from_book}
\title{This function identifies and outputs any variables that appear in the data but are missing in the codebook}
\usage{
missing_from_book(book, dataset)
}
\arguments{
\item{book}{A dataframe containing the GHDx formatted codebook}

\item{dataset}{A dataframe containing the GHDx formatted dataset}
}
\value{
TRUE if no mismatched variables are identified. False if otherwise. Also prints messages to the log.
}
\description{
This function identifies and outputs any variables that appear in the data but are missing in the codebook
}
