% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghdx_micro_checks.R
\name{missing_from_data}
\alias{missing_from_data}
\title{This function identifies and outputs any variables that appear in the codebook but are missing in the data}
\usage{
missing_from_data(book, dataset)
}
\arguments{
\item{book}{A dataframe containing the GHDx formatted codebook}

\item{dataset}{A dataframe containing the GHDx formatted dataset}
}
\value{
TRUE if no mismatched variables are identified. False if otherwise. Also prints messages to the log.
}
\description{
This function identifies and outputs any variables that appear in the codebook but are missing in the data
}
